Drop num-traits dependency
The crate was used in two places:

* `src/cff.rs` — five generic bounds of the form
  `F: num::Unsigned + Copy` / `N: num::Unsigned + Copy` on
  `Range<F, N>` impls and helpers. The `num::Unsigned` half is purely
  documentary — the function bodies only use the `From<...>` projections
  to `u16`/`u32`/`usize` (which are listed alongside in the where
  clauses). Drop the `num::Unsigned` part of each bound; `Copy` plus the
  conversion bounds are sufficient for compilation.
* `src/bitmap.rs::bigger_or_closer_to_zero` — generic over
  `V: PartialOrd + num::Signed + num::Zero`. The two callers pass an
  `i16` and an `i32`. Specialise the function to take `i32` and have
  the `i16` caller widen with `i32::from(...)`. The whole-number
  widen-then-compare is bit-identical to the previous behaviour and
  removes the trait soup.

Drop `use num_traits as num` lines and the `num-traits` entry in
Cargo.toml. Also removes the `autocfg` build dep that num-traits
pulls in.

+++ src/bitmap/cbdt.rs
@@ -623,7 +623,10 @@ impl<'a> CBLCTable<'a> {
                     best = Some((difference, bitmap_size, index))
                 }
                 Some((current_best_difference, _, _))
-                    if super::bigger_or_closer_to_zero(difference, current_best_difference) =>
+                    if super::bigger_or_closer_to_zero(
+                        i32::from(difference),
+                        i32::from(current_best_difference),
+                    ) =>
                 {
                     best = Some((difference, bitmap_size, index))
                 }
